ispravljanje bagova,....

+++ server/jgspApp/api/controllers/stationController.js
@@ -2,43 +2,39 @@ var mongoose = require('mongoose');
 var Station = mongoose.model('station');
 var Line = mongoose.model('line');
 
-module.exports.deleteStation = function(req, res)
-{
-    
+module.exports.deleteStation = function(req, res){  
     if(!req.params._id ) {
-        sendJSONresponse(res, 400, {
-            "message": "all fields must be filled in"
-        });
-        return;
-
+        return res.status(400).json({ "message": "You must select station from combo box!"})
     }
 
-    Station.findOneAndRemove({_id: req.params._id}).then(bla =>{
+    Station.findOne({_id: req.params._id}).then(data=>{
+        if(!data){
+            return res.status(400).json({"message": "The selected station does not exist in the database!"})
+        }
+        else{
+            Station.findOneAndRemove({_id: req.params._id}).then(bla =>{
 
-        Line.find().then(aa => {
-            aa.forEach(bb => {
-                bb.stations.forEach(cc => {
-                    if(cc == req.params._id)
-                    {
-                        bb.stations.remove(cc);
-                        Line.findOneAndUpdate({_id: bb._id}, {stations:bb.stations}).then(abc => {});
-                    }
+                Line.find().then(aa => {
+                    aa.forEach(bb => {
+                        bb.stations.forEach(cc => {
+                            if(cc == req.params._id) {
+                                bb.stations.remove(cc);
+                                Line.findOneAndUpdate({_id: bb._id}, {stations:bb.stations}).then(abc => {});
+                            }
+                        })
+                    })
                 })
-            })
-        })
-        res.status(200).json({
-            "message" : "Station successfully deleted!"
-    });
-});
-}
+                return res.status(200).json({"message" : "Station successfully deleted!" });
+            });
+        }
+    })
 
+    
+}
 
 module.exports.changeStation = function(req, res){
     if(!req.body.Name || !req.body.AddressStation || !req.body.Latitude || !req.body.Longitude ) {
-        sendJSONresponse(res, 400, {
-            "message": "All fields required"
-        });
-        return;
+        return res.status(400).json({ "message": "You must complete all the fields!"})
     }
 
     Station.findById(req.body.Id).exec().then(st=>{
@@ -49,9 +45,7 @@ module.exports.changeStation = function(req, res){
             var vers = req.body.Version + 1;
             const nest = { address : req.body.Address, name : req.body.Name, latitude : req.body.Latitude, longitude: req.body.Longitude, __v: vers}
             Station.findOneAndUpdate({_id : req.body.Id}, nest).then(bla => {
-                res.status(200).json({
-                "message" : "Station successfully changed!"
-                });
+                res.status(200).json({ "message" : "Station successfully changed!"});
             })
         }
     })
@@ -65,7 +59,7 @@ module.exports.getAllStations = function(req, res){
 
 module.exports.addStation = function(req, res){
     if(!req.body.Name || !req.body.AddressStation || !req.body.Latitude || !req.body.Longitude ) {
-        return res.status(400).json({ "message": "You must complete all the fields!"});
+        return res.status(400).json({ "message": "You must complete all the fields!"})
     }
 
     if((req.body.Name.charAt(0)) >= "0" && (req.body.Name.charAt(0)) <= "9"){
